requirements.txt
# Image processing
Pillow>=10.0.0

# Video processing
ffmpeg-python>=0.2.0

# AWS S3 support
boto3>=1.28.0

# Database ORM and drivers
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.0

# Configuration management
python-dotenv>=1.0.0

# Vector database
pgvector>=0.2.0

# Local Embedding models
sentence-transformers>=2.2.0
transformers>=4.30.0
torch>=2.0.0

# CLIP dependencies
ftfy>=6.1.0
regex>=2023.0.0

# Audio metadata
mutagen>=1.47.0

# Spreadsheet parsing
pandas>=2.0.0
openpyxl>=3.1.0
odfpy>=1.4.1

# Document parsing
python-docx>=1.1.0
PyPDF2>=3.0.0

# Code file analysis
chardet>=5.2.0